+++ Kernel/cpan-lib/Excel/Writer/XLSX/Package/ContentTypes.pm
@@ -21,7 +21,7 @@ use Carp;
 use Excel::Writer::XLSX::Package::XMLwriter;
 
 our @ISA     = qw(Excel::Writer::XLSX::Package::XMLwriter);
-our $VERSION = '0.84';
+our $VERSION = '0.85';
 
 
 ###############################################################################