+++ Kernel/cpan-lib/Excel/Writer/XLSX/Package/Packager.pm
@@ -31,7 +31,7 @@ use Excel::Writer::XLSX::Package::Theme;
 use Excel::Writer::XLSX::Package::VML;
 
 our @ISA     = qw(Exporter);
-our $VERSION = '0.84';
+our $VERSION = '0.85';
 
 
 ###############################################################################